+++ sample/feature/tiles/src/main/kotlin/com/mihaicristiancondrea/android/apps/apptoolkit/feature/tiles/ui/services/CoinFlipTileService.kt
@@ -19,12 +19,11 @@ package com.mihaicristiancondrea.android.apps.apptoolkit.feature.tiles.ui.servic
 
 import android.os.Build
 import android.service.quicksettings.Tile
-import android.service.quicksettings.TileService
 import com.mihaicristiancondrea.android.apps.apptoolkit.feature.tiles.R
 import kotlin.random.Random
 
 /** Quick Settings tile that flips a virtual coin. */
-class CoinFlipTileService : TileService() {
+class CoinFlipTileService : TrackedTileService() {
     override fun onStartListening() {
         super.onStartListening()
         qsTile?.apply {