+++ sample/feature/tiles/src/main/kotlin/com/mihaicristiancondrea/android/apps/apptoolkit/feature/tiles/ui/services/FlashDimmerTileService.kt
@@ -19,7 +19,6 @@ package com.mihaicristiancondrea.android.apps.apptoolkit.feature.tiles.ui.servic
 
 import android.os.Build
 import android.service.quicksettings.Tile
-import android.service.quicksettings.TileService
 import com.mihaicristiancondrea.android.apps.apptoolkit.feature.tiles.R
 import com.mihaicristiancondrea.android.apps.apptoolkit.feature.tiles.data.repositories.MorseRepository
 import com.mihaicristiancondrea.android.apps.apptoolkit.feature.tiles.data.repositories.TorchRepository
@@ -27,7 +26,7 @@ import org.koin.core.component.KoinComponent
 import org.koin.core.component.inject
 
 /** Quick Settings control that cycles through the distinct levels supported by the device. */
-class FlashDimmerTileService : TileService(), KoinComponent {
+class FlashDimmerTileService : TrackedTileService(), KoinComponent {
 
     private val torchRepository: TorchRepository by inject()
     private val morseRepository: MorseRepository by inject()